Store reformation, optimised queries to backend with this reformation
