+++ client/src/components/administration/administration.vue
@@ -4,7 +4,8 @@ import { Disciplines, PanelsList } from "./components/components";
 import { defaultAdministratableOptions } from "./common/common";
 
 import styles from "./styles.module.scss";
-import { ref } from "@/hooks/hooks";
+import { ref, useStore } from "@/hooks/hooks";
+import { AdministrationActions } from "@/store/actions.common";
 
 const ShownPanel: Component = ref<Component>(Disciplines);
 
@@ -14,6 +15,15 @@ const handlePanelChoose: (id: number) => () => void =
       ({ id: optionId }) => optionId === id
     )?.component as Component;
   };
+
+const store = useStore();
+
+store.dispatch(AdministrationActions.GET_ALL_DISICPLINES);
+store.dispatch(AdministrationActions.GET_ALL_FACULTIES);
+store.dispatch(AdministrationActions.GET_ALL_TEACHERS);
+store.dispatch(AdministrationActions.GET_ALL_DEPARTMENTS);
+store.dispatch(AdministrationActions.GET_ALL_GROUPS);
+store.dispatch(AdministrationActions.GET_ALL_STUDENTS);
 </script>
 
 <template>
