+++ client/src/components/administration/components/faculties/components/faculties-cards-list/components/faculty-card/components/update-faculty-form/update-faculty-form.vue
@@ -7,7 +7,7 @@ import type {
 import { reactive, useStore } from "@/hooks/hooks";
 import { faculty as facultyValidator } from "@/validators/validators";
 import type { ValidationError } from "@/exceptions/exceptions";
-import { FacultiesActions } from "@/store/actions.common";
+import { AdministrationActions } from "@/store/actions.common";
 
 import styles from "./styles.module.scss";
 
@@ -62,7 +62,7 @@ const handleSubmit: (event: Event) => void = (event: Event) => {
   if (
     Object.values(facultyUpdateValidationState).every((el) => el.length === 0)
   ) {
-    store.dispatch(FacultiesActions.UPDATE_FACULTY, {
+    store.dispatch(AdministrationActions.UPDATE_FACULTY, {
       id: props.initialFaculty.id,
       payload: facultyUpdateFormState,
     });
