+++ client/src/components/administration/components/students/students.vue
@@ -1,24 +1,20 @@
 <script lang="ts" setup>
 import { CreateStudentForm, StudentsCardsList } from "./components/components";
 import { computed, ref, useStore } from "@/hooks/hooks";
-import { StudentsActions, GroupsActions } from "@/store/actions.common";
 import { Button } from "@/common/components/components";
 
 import styles from "./styles.module.scss";
 
 const store = useStore();
-const students = computed(() => store.state.students.students);
+const students = computed(() => store.state.administration.students);
 const isStudentCrеationPossible = computed(
-  () => store.state.groups.groups.length > 0
+  () => store.state.administration.groups.length > 0
 );
 
 const studentCreationFormShowState = ref<boolean>(false);
 const handleToggle: () => void = (): void => {
   studentCreationFormShowState.value = !studentCreationFormShowState.value;
 };
-
-store.dispatch(GroupsActions.GET_ALL_GROUPS);
-store.dispatch(StudentsActions.GET_ALL_STUDENTS);
 </script>
 
 <template>
