+++ client/src/store/actions.common.ts
@@ -1,7 +1,1 @@
-export { Actions as DepartmentsActions } from "./departments/actions.common";
-export { Actions as DisciplinesActions } from "./disciplines/actions.common";
-export { Actions as FacultiesActions } from "./faculties/actions.common";
-export { Actions as GroupsActions } from "./groups/actions.common";
-export { Actions as SchedulesActions } from "./schedules/actions.common";
-export { Actions as StudentsActions } from "./students/actions.common";
-export { Actions as TeachersActions } from "./teachers/actions.common";
+export { Actions as AdministrationActions } from "./administration/actions.common";
