+++ client/src/store/administration/actions.common.ts
@@ -1,0 +1,32 @@
+enum Actions {
+  GET_ALL_DEPARTMENTS = "administration/getAllDepartments",
+  CREATE_DEPARTMENT = "administration/createDepartment",
+  UPDATE_DEPARTMENT = "administration/updateDepartment",
+  DELETE_DEPARTMENT = "administration/deleteDepartment",
+  GET_ALL_DISICPLINES = "administration/getAllDisciplines",
+  CREATE_DISCIPLINE = "administration/createDiscipline",
+  UPDATE_DISCIPLINE = "administration/updateDiscipline",
+  DELETE_DISCIPLINE = "administration/deleteDiscipline",
+  GET_ALL_FACULTIES = "administration/getAllFaculties",
+  CREATE_FACULTY = "administration/createFaculty",
+  UPDATE_FACULTY = "administration/updateFaculty",
+  DELETE_FACULTY = "administration/deleteFaculty",
+  GET_ALL_GROUPS = "administration/getAllGroups",
+  CREATE_GROUP = "administration/createGroup",
+  UPDATE_GROUP = "administration/updateGroup",
+  DELETE_GROUP = "administration/deleteGroup",
+  GET_ALL_SCHEDULES = "administration/getAllSchedules",
+  CREATE_SCHEDULE = "administration/createSchedule",
+  UPDATE_SCHEDULE = "administration/updateSchedule",
+  DELETE_SCHEDULE = "administration/deleteSchedule",
+  GET_ALL_STUDENTS = "administration/getAllStudents",
+  CREATE_STUDENT = "administration/createStudent",
+  UPDATE_STUDENT = "administration/updateStudent",
+  DELETE_STUDENT = "administration/deleteStudent",
+  GET_ALL_TEACHERS = "administration/getAllTeachers",
+  CREATE_TEACHER = "administration/createTeacher",
+  UPDATE_TEACHER = "administration/updateTeacher",
+  DELETE_TEACHER = "administration/deleteTeacher",
+}
+
+export { Actions };
