--- client/src/store/departments/actions.common.ts
@@ -1,8 +1,0 @@
-enum Actions {
-  GET_ALL_DEPARTMENTS = "departments/getAllDepartments",
-  CREATE_DEPARTMENT = "departments/createDepartment",
-  UPDATE_DEPARTMENT = "departments/updateDepartment",
-  DELETE_DEPARTMENT = "departments/deleteDepartment",
-}
-
-export { Actions };
