--- client/src/store/departments/state.ts
@@ -1,9 +1,0 @@
-import type { DataStatus } from "@/common/enums/enums";
-import type { DepartmentsGetAllItemResponseDto } from "@/common/types/types";
-
-type State = {
-  departments: DepartmentsGetAllItemResponseDto[];
-  dataStatus: DataStatus;
-};
-
-export { type State };
