--- client/src/store/disciplines/actions.common.ts
@@ -1,8 +1,0 @@
-enum Actions {
-  GET_ALL_DISICPLINES = "disciplines/getAllDisciplines",
-  CREATE_DISCIPLINE = "disciplines/createDiscipline",
-  UPDATE_DISCIPLINE = "disciplines/updateDiscipline",
-  DELETE_DISCIPLINE = "disciplines/deleteDiscipline",
-}
-
-export { Actions };
