--- client/src/store/disciplines/state.ts
@@ -1,9 +1,0 @@
-import type { DataStatus } from "@/common/enums/enums";
-import type { DisciplinesGetAllItemResponseDto } from "@/common/types/types";
-
-type State = {
-  disciplines: DisciplinesGetAllItemResponseDto[];
-  dataStatus: DataStatus;
-};
-
-export { type State };
