--- client/src/store/faculties/actions.common.ts
@@ -1,8 +1,0 @@
-enum Actions {
-  GET_ALL_FACULTIES = "faculties/getAllFaculties",
-  CREATE_FACULTY = "faculties/createFaculty",
-  UPDATE_FACULTY = "faculties/updateFaculty",
-  DELETE_FACULTY = "faculties/deleteFaculty",
-}
-
-export { Actions };
