--- client/src/store/groups/actions.common.ts
@@ -1,8 +1,0 @@
-enum Actions {
-  GET_ALL_GROUPS = "groups/getAllGroups",
-  CREATE_GROUP = "groups/createGroup",
-  UPDATE_GROUP = "groups/updateGroup",
-  DELETE_GROUP = "groups/deleteGroup",
-}
-
-export { Actions };
