--- client/src/store/schedules/actions.common.ts
@@ -1,8 +1,0 @@
-enum Actions {
-  GET_ALL_SCHEDULES = "schedules/getAllSchedules",
-  CREATE_SCHEDULE = "schedules/createSchedule",
-  UPDATE_SCHEDULE = "schedules/updateSchedule",
-  DELETE_SCHEDULE = "schedules/deleteSchedule",
-}
-
-export { Actions };
