--- client/src/store/students/actions.common.ts
@@ -1,8 +1,0 @@
-enum Actions {
-  GET_ALL_STUDENTS = "students/getAllStudents",
-  CREATE_STUDENT = "students/createStudent",
-  UPDATE_STUDENT = "students/updateStudent",
-  DELETE_STUDENT = "students/deleteStudent",
-}
-
-export { Actions };
