--- client/src/store/students/state.ts
@@ -1,9 +1,0 @@
-import type { DataStatus } from "@/common/enums/enums";
-import type { StudentsGetAllItemResponseDto } from "@/common/types/types";
-
-type State = {
-  students: StudentsGetAllItemResponseDto[];
-  dataStatus: DataStatus;
-};
-
-export { type State };
