--- client/src/store/teachers/actions.common.ts
@@ -1,8 +1,0 @@
-enum Actions {
-  GET_ALL_TEACHERS = "teachers/getAllTeachers",
-  CREATE_TEACHER = "teachers/createTeacher",
-  UPDATE_TEACHER = "teachers/updateTeacher",
-  DELETE_TEACHER = "teachers/deleteTeacher",
-}
-
-export { Actions };
