--- client/src/store/teachers/state.ts
@@ -1,9 +1,0 @@
-import type { DataStatus } from "@/common/enums/enums";
-import type { TeachersGetAllItemResponseDto } from "@/common/types/types";
-
-type State = {
-  teachers: TeachersGetAllItemResponseDto[];
-  dataStatus: DataStatus;
-};
-
-export { type State };
